[GITHUB-13] Fixes purge on Bionic
Purge was hanging on interactive dialog during an apt operation,
adding the noninteractive var fixes this issue.

Also updated the Linux Purge script to the latest version from
the author and stopped it from running in Docker as this fails
without a reboot, which also doesn't work in Docker :)

Finally pegged Molecule to a version which still supports
Ansible 2.3.